# projekat/laravelDomaci/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Event Management API",
        "description": "API documentation for Event Management System with Laravel and React",
        "contact": {
            "email": "admin@eventmanagement.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/api/tickets/purchase": {
            "post": {
                "tags": [
                    "Tickets"
                ],
                "summary": "Purchase a ticket for an event",
                "operationId": "20c50efe26f3068218a5548c98385c83",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "event_id"
                                ],
                                "properties": {
                                    "event_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "discount_percentage": {
                                        "type": "number",
                                        "example": 10
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Ticket purchased successfully"
                    },
                    "422": {
                        "description": "Validation failed or no tickets available"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/tickets/my": {
            "get": {
                "tags": [
                    "Tickets"
                ],
                "summary": "Get tickets for the authenticated user",
                "operationId": "76d81fa3bcbe69bcc5a06a4dbddfa826",
                "responses": {
                    "200": {
                        "description": "List of user's tickets"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/tickets/{id}": {
            "get": {
                "tags": [
                    "Tickets"
                ],
                "summary": "Get a single ticket by ID",
                "operationId": "da5292e25da44db69ed981a43357753c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ticket details"
                    },
                    "403": {
                        "description": "Unauthorized access"
                    },
                    "404": {
                        "description": "Ticket not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/tickets/{id}/cancel": {
            "put": {
                "tags": [
                    "Tickets"
                ],
                "summary": "Cancel a ticket",
                "operationId": "0bd6107fe0b09f4514f2ec73ef52793a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ticket cancelled"
                    },
                    "422": {
                        "description": "Cannot cancel ticket"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/tickets/validate/{ticketNumber}": {
            "get": {
                "tags": [
                    "Tickets"
                ],
                "summary": "Validate a ticket by ticket number",
                "operationId": "c535eabbb718a22a7d226c5bfebfbdb7",
                "parameters": [
                    {
                        "name": "ticketNumber",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ticket is valid"
                    },
                    "404": {
                        "description": "Invalid ticket number"
                    }
                }
            }
        },
        "/api/tickets/{id}/mark-used": {
            "put": {
                "tags": [
                    "Tickets"
                ],
                "summary": "Mark a ticket as used",
                "operationId": "f3d6ac4a6e85126ae3fad1fbdcd12175",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ticket marked as used"
                    },
                    "422": {
                        "description": "Ticket not active"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}